Ensure value of DefaultPromise can not leak when EventExecutor is shutdown.
Motivation:

If the EventExecutor is shutdown from which a DefaultPromise was created before someone completed the promise, it will not be able to schedule the notification of the attached listeners. This can lead to leaks as the user may expect that its possible to release the passed in value (which may be ReferenceCounted) in a listener.

Modifications:

The modifications needed to make this 100 % safe are much more complicated then I hoped for. The problem is we really need to ensure we can schedule the notification of the listeners before we set the result that is visible by the user as otherwise the ownership of the value (which may be ReferenceCounted) is not clear and so its not clear who is responsible in releasing it. Because of this we need to use a "two-step" process when setting the value when we schedule the notification of the listeners via the EventExecutor.

Result:

Fixes https://github.com/netty/netty/issues/8007.